requirements.txt
PyGObject>=3.42.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
